+++ js/config/configs/react-with-hooks-eslintrc.js
@@ -3,7 +3,7 @@
 
 module.exports = {
   extends: './configs/react-eslintrc.js',
-  plugins: ['react-hooks'],
+  plugins: [ 'react-hooks' ],
   rules: {
     'react-hooks/rules-of-hooks': 'error',
     'react-hooks/exhaustive-deps': 'error'